#1363 Refactor DataSourceParameterExtractor & Add HikariDataSourceParameterExtractor.
