+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/extractor/DBCPDataSourceParameterExtractor.java
@@ -26,7 +26,7 @@ import javax.sql.DataSource;
  *
  * @author zhaojun
  */
-public final class DBCPDataSourceParameterExtractor extends DefaultDataSourceParameterExtractor {
+public final class DBCPDataSourceParameterExtractor extends DataSourceParameterExtractorAdapter {
     
     DBCPDataSourceParameterExtractor(final DataSource dataSource) {
         super(dataSource);
