+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/extractor/DruidDataSourceParameterExtractor.java
@@ -26,7 +26,7 @@ import javax.sql.DataSource;
  *
  * @author zhaojun
  */
-public final class DruidDataSourceParameterExtractor extends DefaultDataSourceParameterExtractor {
+public final class DruidDataSourceParameterExtractor extends DataSourceParameterExtractorAdapter {
     
     DruidDataSourceParameterExtractor(final DataSource dataSource) {
         super(dataSource);
